+++ ecs/paged.go
@@ -2,19 +2,27 @@ package ecs
 
 const fixedPageSize = 32
 
-// pagedArr32 is a paged collection working with pages of length 32 arrays.
+// pagedSlice is a paged collection working with pages of length 32 slices.
+// It's primary purpose is pointer persistence, which is not given using simple slices.
 //
 // Implements [archetypes].
-type pagedArr32[T any] struct {
-	pages   [][fixedPageSize]T
-	len     int
-	lenLast int
+type pagedSlice[T any] struct {
+	pages    [][]T
+	len      int
+	lenLast  int
+	pageSize int
+}
+
+func newPagedSlice[T any](pageSize int) pagedSlice[T] {
+	return pagedSlice[T]{
+		pageSize: pageSize,
+	}
 }
 
 // Add adds a value to the paged array.
-func (p *pagedArr32[T]) Add(value T) {
+func (p *pagedSlice[T]) Add(value T) {
 	if p.len == 0 || p.lenLast == fixedPageSize {
-		p.pages = append(p.pages, [fixedPageSize]T{})
+		p.pages = append(p.pages, make([]T, fixedPageSize))
 		p.lenLast = 0
 	}
 	p.pages[len(p.pages)-1][p.lenLast] = value
@@ -23,12 +31,12 @@ func (p *pagedArr32[T]) Add(value T) {
 }
 
 // Get returns the value at the given index.
-func (p *pagedArr32[T]) Get(index int) *T {
+func (p *pagedSlice[T]) Get(index int) *T {
 	return &p.pages[index/fixedPageSize][index%fixedPageSize]
 }
 
 // Len returns the current number of items in the paged array.
-func (p *pagedArr32[T]) Len() int {
+func (p *pagedSlice[T]) Len() int {
 	return p.len
 }
 
@@ -36,7 +44,7 @@ func (p *pagedArr32[T]) Len() int {
 //
 // Implements [archetypes].
 type pagedPointerArr32[T any] struct {
-	pages   [][fixedPageSize]*T
+	pages   [][]*T
 	len     int
 	lenLast int
 }
@@ -44,7 +52,7 @@ type pagedPointerArr32[T any] struct {
 // Add adds a value to the paged array.
 func (p *pagedPointerArr32[T]) Add(value *T) {
 	if p.len == 0 || p.lenLast == fixedPageSize {
-		p.pages = append(p.pages, [fixedPageSize]*T{})
+		p.pages = append(p.pages, make([]*T, fixedPageSize))
 		p.lenLast = 0
 	}
 	p.pages[len(p.pages)-1][p.lenLast] = value